pyproject.toml
[tool.poetry]
name = "litemq"
version = "0.1.0"
description = ""
authors = ["thekashifmalik <emailkashifmalik@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8,<4"
grpcio = "^1.67.1"
grpcio-tools = "^1.67.1"
grpclib = "^0.4.7"
protobuf = "^5.28.3"
protoletariat = "^3.3.2"
pytest = "^8.3.4"
pytest-asyncio = "^0.24.0"

[tool.pytest.ini_options]
asyncio_mode = "auto"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
